Attempt to get Ontology resource reasoning working in a limited sense.
Previously we were asking if a resource has any relations and processing
them in one query. This is fast, but prevents inference working since we
weren't explicitly specifying the relation type (e.g. skos:narrower).
Now we hardcode a whole load of potential resource relations and
query explicitly for each one, which is incredibly slow and inefficient
but allows inverseOf relationships to be inferred where relevant.